requirements.txt
baidu-aip==2.2.11.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
colorama==0.4.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
dill==0.2.8.2
idna==2.8
kiwisolver==1.0.1
lxml==4.3.0
matplotlib==3.0.2
multiprocess==0.70.6.1
networkx==2.2
numpy==1.16.0
opencv-python==4.0.0.21
Pillow==5.4.1
prettytable==0.7.2
PyJWT==1.7.1
pyparsing==2.3.1
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
requests==2.21.0
scikit-image==0.14.1
scipy==1.2.0
six==1.12.0
threadpool==1.3.2
toolz==0.9.0
twilio==6.23.1
urllib3==1.24.1
